+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Remove bintray plugin support
+* In multi-module projects plugin use root project for defaults (and license and changelog files resolution) 
 
 ### 1.2.0 (2020-01-19)
 * Remove ".git" suffix for default scm urls